phase-3/requirements.txt
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
elastic-transport==8.10.0
elasticsearch==8.11.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
ndjson==0.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
